Add files via upload
The following is a python script that creates an updated fantasy ranking spreadsheet on Excel. 

It is divided by position (running back, quarter back, wide receiver, tight end, kicker and defense)

The rankings were taken from different websites and displayed on the Excel tab with the player names and teams they play for.

NOTE: to run this, need to have the "BeautifulSoup" library along with "xlsxwriter" library on python. Also need to change line 194 to your desired file location.